--- src/main/java/team_alcoholic/jumo_server/common/domain/BaseEntity.java
@@ -1,18 +1,0 @@
-package team_alcoholic.jumo_server.common.domain;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.MappedSuperclass;
-import org.springframework.data.annotation.CreatedBy;
-import org.springframework.data.annotation.LastModifiedBy;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-@MappedSuperclass
-public class BaseEntity extends BaseTimeEntity {
-    @CreatedBy
-    @Column(updatable = false)
-    private String createdBy;
-
-    @LastModifiedBy
-    private String lastModifiedBy;
-}
