--- src/main/java/team_alcoholic/jumo_server/config/CustomAuthenticationEntryPoint.java
@@ -1,31 +1,0 @@
-package team_alcoholic.jumo_server.config;
-
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.AuthenticationEntryPoint;
-import org.springframework.stereotype.Component;
-
-import java.io.IOException;
-import java.util.HashMap;
-import java.util.Map;
-
-@Component
-public class CustomAuthenticationEntryPoint implements AuthenticationEntryPoint {
-
-    private final ObjectMapper objectMapper = new ObjectMapper();
-
-    @Override
-    public void commence(HttpServletRequest request, HttpServletResponse response, AuthenticationException authException) throws IOException {
-        response.setContentType("application/json;charset=UTF-8");
-        response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
-        Map<String, String> errorDetails = new HashMap<>();
-        errorDetails.put("error", "Unauthorized");
-        errorDetails.put("message", "로그인이 필요합니다.");
-
-        response.getWriter().write(objectMapper.writeValueAsString(errorDetails));
-    }
-}
