--- src/main/java/team_alcoholic/jumo_server/controller/MainController.java
@@ -1,13 +1,0 @@
-package team_alcoholic.jumo_server.controller;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class MainController {
-
-    @GetMapping("/")
-    public String hello() {
-        return "Hello, World!";
-    }
-}
