+++ src/main/java/team_alcoholic/jumo_server/domain/meeting/dto/MeetingListResponseDto.java
@@ -1,0 +1,16 @@
+package team_alcoholic.jumo_server.domain.meeting.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+import java.util.List;
+
+@Getter
+@Setter
+@AllArgsConstructor
+public class MeetingListResponseDto {
+    private List<MeetingListDto> meetings;
+    private Long lastId;
+    private boolean eof;
+}
