+++ src/main/java/team_alcoholic/jumo_server/domain/meeting/repository/MeetingRepository.java
@@ -1,0 +1,12 @@
+package team_alcoholic.jumo_server.domain.meeting.repository;
+
+import team_alcoholic.jumo_server.domain.meeting.dto.MeetingDto;
+import team_alcoholic.jumo_server.domain.meeting.dto.MeetingListDto;
+
+import java.util.List;
+
+public interface MeetingRepository {
+    MeetingDto findMeetingById(Long id);
+    List<MeetingListDto> findLatestMeetingList(int limit);
+    List<MeetingListDto> findLatestMeetingListById(int limit, Long cursor);
+}
