+++ src/main/java/team_alcoholic/jumo_server/domain/region/repository/RegionRepository.java
@@ -1,0 +1,10 @@
+package team_alcoholic.jumo_server.domain.region.repository;
+
+import team_alcoholic.jumo_server.domain.region.domain.Region;
+
+import java.util.List;
+
+public interface RegionRepository {
+    Region findByAdmcd(String admcd);
+    List<Region> findByAdmcdLike(String admcd);
+}
