+++ src/main/java/team_alcoholic/jumo_server/global/common/domain/BaseEntity.java
@@ -1,0 +1,16 @@
+package team_alcoholic.jumo_server.global.common.domain;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.MappedSuperclass;
+import org.springframework.data.annotation.CreatedBy;
+import org.springframework.data.annotation.LastModifiedBy;
+
+@MappedSuperclass
+public class BaseEntity extends BaseTimeEntity {
+    @CreatedBy
+    @Column(updatable = false)
+    private String createdBy;
+
+    @LastModifiedBy
+    private String lastModifiedBy;
+}
