--- src/main/java/team_alcoholic/jumo_server/region/domain/Region.java
@@ -1,15 +1,0 @@
-package team_alcoholic.jumo_server.region.domain;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import lombok.Getter;
-import team_alcoholic.jumo_server.common.domain.BaseTimeEntity;
-
-@Entity
-@Getter
-public class Region extends BaseTimeEntity {
-
-    @Id
-    private String admcd;
-    private String admnm;
-}
