--- src/main/java/team_alcoholic/jumo_server/region/repository/RegionRepository.java
@@ -1,10 +1,0 @@
-package team_alcoholic.jumo_server.region.repository;
-
-import team_alcoholic.jumo_server.region.domain.Region;
-
-import java.util.List;
-
-public interface RegionRepository {
-    Region findByAdmcd(String admcd);
-    List<Region> findByAdmcdLike(String admcd);
-}
