--- src/main/java/team_alcoholic/jumo_server/user/dto/UserDTO.java
@@ -1,20 +1,0 @@
-package team_alcoholic.jumo_server.user.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
-
-/**
- * 사용자 정보를 담는 DTO
- */
-@Getter
-@Setter
-@Builder
-public class UserDTO {
-    private Long id;
-    private String provider;
-    private String providerId;
-    private String profileNickname;
-    private String profileImage;
-    private String profileThumbnailImage;
-}
